Move util.c to objective_c_helper, update imports (#2272)
