+++ pkgs/objective_c/hook/build.dart
@@ -15,7 +15,6 @@ const assetName = 'objective_c.dylib';
 
 // TODO(https://github.com/dart-lang/native/issues/2272): Remove this from the
 // main build.
-const testFiles = ['test/util.c'];
 
 final logger = Logger('')
   ..level = Level.INFO
@@ -62,11 +61,6 @@ void main(List<String> args) async {
     // Only include the test utils on mac OS. They use memory functions that
     // aren't supported on iOS, like mach_vm_region. We don't need them on iOS
     // anyway since we only run memory tests on mac.
-    if (os == OS.macOS) {
-      cFiles.addAll(
-        testFiles.map((f) => input.packageRoot.resolve(f).toFilePath()),
-      );
-    }
 
     final sysroot = sdkPath(codeConfig);
     final minVersion = minOSVersion(codeConfig);
