+++ pkgs/objective_c/test/autorelease_test.dart
@@ -11,7 +11,7 @@ import 'dart:ffi';
 import 'package:objective_c/objective_c.dart';
 import 'package:test/test.dart';
 
-import 'util.dart';
+import 'package:objective_c_helper/src/util.dart';
 
 void main() {
   group('autoReleasePool', () {
