+++ pkgs/objective_c/test/interface_lists_test.dart
@@ -12,7 +12,7 @@ import 'package:ffigen/src/code_generator/objc_built_in_types.dart';
 import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
 
-import 'util.dart';
+import 'package:objective_c_helper/src/util.dart';
 
 // The default expect error message for sets isn't very useful. In the common
 // case where the sets are different lengths, the default matcher just says the
